docs: convert documentation to texinfo
Migrates the existing documentation into a texinfo manual. The benefit
of using texinfo is that it allows for typesetting and portability.

The typesetting capabilities of texinfo allows for setting variables
in the document as well as enable more structured editing and writing
of new sections.

The increased portability comes from the fact that texinfo can be read
in a variety of formats: ranging from plaintext to web. It can also be
interactively read on a system using `info`