flake: add manual package; integrate into installer
This integrates the documentation into the installer as a package. It
exposes two scripts: "start-manual-info" and "start-manual-web" which
are able to open the documentation on the installer ISO to be used as
needed (for example, opening the info manual on TTY2).

This makes the installation guide and manual more portable as it can
be accessed both on a minimal ISO and on a host system.