flake: add "apps" output to allow for viewing the manual
Solves being able to build the manual on all nix supported systems
without having to go through cross on the packages output. This allows
darwin users to view the manual.

The `apps` output of a flake corresponds to the `nix run` command so
this allows for building and viewing the manual just through `nix run`
and there's a start script specifically created for darwin
users (which replaces `xdg-open` with `open`).