docs/manual: set "next" manual; add instructions for viewing manuals
Adds a "-next" suffix to the version of the manual to distinguish it
from release versions of manuals. This tackles the issues of upstream
documentation moving ahead of what's in the versioned release.

Also noted this in the release process section of the manual. This
way, the documentation can stay updated and current with users who
follow the development branch.

This also helps maintainers incrementally add changes per commit to
the release notes instead of having to summarize at the end.